chore: rename to BinaryCataloger
