+++ syft/pkg/cataloger/binary/cataloger_test.go
@@ -1,4 +1,4 @@
-package file
+package binary
 
 import (
 	"testing"
@@ -94,7 +94,7 @@ func TestClassifierCataloger_DefaultClassifiers_PositiveCases(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			c := NewFileCataloger()
+			c := NewBinaryCataloger()
 
 			src, err := source.NewFromDirectory(test.fixtureDir)
 			require.NoError(t, err)
@@ -144,7 +144,7 @@ func TestClassifierCataloger_DefaultClassifiers_PositiveCases_Image(t *testing.T
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			c := NewFileCataloger()
+			c := NewBinaryCataloger()
 
 			img := imagetest.GetFixtureImage(t, "docker-archive", test.fixtureImage)
 			src, err := source.NewFromImage(img, "test-img")
@@ -173,7 +173,7 @@ func TestClassifierCataloger_DefaultClassifiers_PositiveCases_Image(t *testing.T
 }
 
 func TestClassifierCataloger_DefaultClassifiers_NegativeCases(t *testing.T) {
-	c := NewFileCataloger()
+	c := NewBinaryCataloger()
 
 	src, err := source.NewFromDirectory("test-fixtures/classifiers/negative")
 	assert.NoError(t, err)
