+++ syft/pkg/cataloger/binary/classifier.go
@@ -1,4 +1,4 @@
-package file
+package binary
 
 import (
 	"bytes"
