+++ syft/pkg/cataloger/binary/classifier_test.go
@@ -1,4 +1,4 @@
-package file
+package binary
 
 import (
 	"testing"
