+++ syft/pkg/cataloger/binary/default_classifiers.go
@@ -1,4 +1,4 @@
-package file
+package binary
 
 import "github.com/anchore/syft/syft/pkg"
