+++ syft/pkg/cataloger/binary/test-fixtures/classifiers/negative/busybox
