+++ syft/pkg/cataloger/binary/test-fixtures/classifiers/negative/libpython2.7.so
