+++ syft/pkg/cataloger/binary/test-fixtures/classifiers/negative/python2.6
