+++ syft/pkg/cataloger/binary/test-fixtures/classifiers/positive/[
