+++ syft/pkg/cataloger/binary/test-fixtures/classifiers/positive/libpython3.7.so
