+++ syft/pkg/cataloger/binary/test-fixtures/classifiers/positive/python3.6
