+++ syft/pkg/cataloger/binary/test-fixtures/image-busybox/Dockerfile
