+++ syft/pkg/cataloger/binary/test-fixtures/version.txt
